Fix crash when file has a transform and a syntax error
`parseDeps` returns undefined on a syntax error, so `deps.concat`ing it
does not work. This patch early returns from `getDeps` if `deps` is
undefined.

This also changes `_transformDeps` to be initialised lazily once the
first 'dep' event happens, so that we're not unnecessarily concating
when there are no additional deps.