Also emit 'file' event for files with syntax errors
Previously, `parseDeps` emitted an error event when there was a syntax
error, and then returned undefined. Inside `persistentCacheFallback()`,
then, the callback would not be called. This used to be OK but the
persistent cache changes in v6 changed the callback, so now _it_ is
responsible for 'file' and 'error' events.

This patch changes `parseDeps` to throw instead of emit an error. This
bubbles up to `persistentCacheFallback()` where it is caught and passed
to the callback. Then that callback will first emit a 'file' event for
the file and only then emit the 'error' event.

This makes watchify aware of the file, so it will watch it and rebuild
when it changes. Fixes #144 (really this time!)