Propagate transform 'file' events
Currently, transforms can unofficially emit file events that are picked
up by watchify. Watchify can then watch those files. This is useful for
transforms that include some external file content, like sheetify with
the `css('module')` syntax.

When using the persistentCache option, 'file' events emitted from
transforms are not cached. So, if a transformed file is loaded from the
persistent cache, related files that were added by transforms will not
be watched by watchify.

This patch stores 'file' events emitted by transforms in an array, which
is passed to the persistent cache handler. When loading from cache,
'file' events are emitted on the module-deps instance for each of the file
paths in that array.

For consistency, (and because it was easier to implement), 'file' events
emitted by transforms are also re-emitted on the module-deps instance
when a file was not loaded from cache. This means watchify doesn't have
to do the `.on('transform', (t) => t.on('file'))` dance anymore. All
relevant files are emitted on the global 'file' event.

This should probably be treated as a breaking change because it is
touching some fickle machinery. People are relying on the 'file' and
'transform'»'file' events in very specific ways.